Revert & Rename
